this new code will detect the direction of the message not by the first strong char, but by the most RTL chars on the line.
i also take care of message in which you first address a person (LTR
string) at the beginning of the line. IMHO some more work should be put
on the direction detection, but that's good enough for now.

see this image to get an idea of how it works now:
http://img238.imageshack.us/my.php?image=koversationbidibug3rm8.jpg


CCMAIL: kde-il@yahoogroups.com
GUI

svn path=/trunk/extragear/network/konversation/; revision=586839