Merge of auto-away support along with related cleanups:
* Added per-identity support for automatic away after a configurable
  duration of user desktop inactivity and/or screensaver activation,
  along with support for automatic return on activity.
* Fixed the "Global Away" toggle to make sense and update its state
  properly.
* Added an '/aunaway' command to complement '/aaway' and '/unaway'.
* Broadly rewrote away management related code for improved robust-
  ness and less inane duplication and hacks (e.g. no more abuse of
  multiServerCommand parsing tricks for global away).
* Some layout improvements in the identity editor dialog.
* Made use of units in spinboxen in the identity and preferences UI
  more consistent.

The copious amount of debug output will be removed a little later.

As a side note, aside from a yet-missing warning dialog that needs
to be added, this concludes active feature development for Konver-
sation 1.1 and marks the beginning of feature freeze. A few tweaks
here and there down the road string freeze will begin, and lead to
a release.

This release cycle has been very long - yes, too long - due to the
usual collection of real life issues and similar, but at the same
time also has a lot to show for it - the longest changelog yet (al-
most a third of the total changelog file), and large amounts of
both under the hood and user-visible improvements, fixes and feat-
ures. This is a much better app than 1.0.1 all around.

Konversation 1.1 will be the last KDE 3-based Konversation. The
next milestone for us will be a KDE 4 release. This will initially
be a fairly straight port for timing reasons (don't want to make
you wait even longer), but rest assured that our plans for future
versions are filled to the brim with high-flying changes and exci-
ting new stuff. All the while making sure that Konversation stays
the friendly, no-nonsense, highly KDE-integrated IRC client you've
come to know, and that continues to serve an important role in
keeping our community of users and developers together.

svn path=/branches/extragear/kde3/network/konversation/; revision=814915