Connect up RPL_LISTEND to notify channellistpanel when it's at the end of the list.
Convert channellistpanel to a model/view approach using qsortfilterproxy/qabstractlistmodel
On refresh we create a new ChannelListModel (qabstractlistmodel) and load it up, then attach it to the qsortfilterproxy, instead of redoing the existing ChannelListModel.
This saves LOTS of time and ugliness. and results in a very fast refresh.

Since qsortfilterproxy seems to filter pretty fast I chose to get rid of the applyFilter button and attach all filter changes to a 300ms singleshot timer.

Change it so we disable the refreshlist button onClick instead of on firstRun to avoid multiple refresh triggers, but also put it on a 3000ms timer to re-enable.
On addToChannelList's firstRun the timer gets disabled if active.
This allows for a server load error or other problem.
If the list happens to get sent after the timer is up, addToChannelList will disable the button again and all is well.

Also, change the topicLabel's channel context menu to mimic the changes to the ircview's.

svn path=/trunk/extragear/network/konversation/; revision=999155