Fix the number of mode chars sent in op/deop
Previously the modes portion would fill up for all the nicks
so if you had 5 nicks, that would split into two lines with 3
on the first and 2 on the second. But on both you would have 5
mode chars. Now we add one mode per nick at the same time the
nick is added.
BUG:214258