Fix auto-who responses becoming visible after parameterless '/who.'
Fixes a bug causing the server responses to background data gather-
ing via WHO (to keep the app's idea of its own hostmask up to date,
as well as optionally channel user info) to be displayed in tabs as
if the user had made the requests manually after sending a '/who'
command without parameters.

For background, 76f47cde1 introduced a change adding a parameterless
'/who' as a 'WHO *' to InputFilter's WHO request tracking list,
claiming that servers interpret a parameterless WHO as a 'WHO *',
which doesn't hold up in reality. While RFC 1459 and RFC 2812 claim
the parameter to be optional, the servers I have tested with don't
allow it in practice, presumably for load reasons: ircd 2.11.1p1 and
ircd-seven 1.0.1 reply with a 461 with "Not enough parameters" and
bahamut 1.8(0.3) with a set of 334's providing a human-readable
usage description of the WHO command. Thus the behavior of 'WHO' and
'WHO *' differ in reality, and this commit changes things so that a
parameterless '/who' no longer results in a 'WHO *' being put on In-
putFilter'S WHO request tracking list.
BUG:227795