Do not cut notifications
Blindly cutting notifications every 50 characters looks bad because the cut
often lands in the middle of the word. It also occasionally lands in the
middle of an html tag (for example because the text contains a '<' or a
'>' character) causing even weirder renderings.

It's better to leave formatting to the notification system.