In-progress automatic reconnect can now be aborted using '/disconnect'.
BUG:235186

A manual disconnect also didn't use to reset the reconnection counter
to 0 - only a successful connect would. Since there was no way to abort
trying to connect until now that wasn't of consequence, but had to be
and is fixed now. Also cleaned up Server::disconnect() shadowing the
QObject version.