Bring tab bar code in line with our new minimum version deps.
Back in the day we tried to set our tab widget to document mode, but
Oxygen at the time was broken with it as it was released prior to the
Qt version introducing document mode. We briefly added a workaround
that made it work with Oxygen, but broke other styles, so we gave up
on the experiment. Now we depend on a KDE version with an Oxygen that
handles document mode properly, so use it.

Also port away from deprecated KTabWidget API now that our mininum re-
quired Qt version offers the new one.