Make search boxes more kde-widget-like
The Channel List and URL Catcher use QSortFilterProxyModels.
KDE has a specific widget to be used for the search box
for these models. However, there are some annoying limitations.
With the URL Catcher it was a simple to adapt to, but the
Channel List is impossible to adapt because it uses regex
which the KFliterProxySearchLine doesn't allow for. So
instead I've chosen to make it look just like the kde widget
so we have some consistency.