Much improve the /cycle command.
Prior to this commit, /cycle only worked from within a particular
channel tab to close it and subsequently rejoin that channel. It's
quite a bit more capable now; the new syntax is:

/CYCLE [-APP | -SERVER] [channel | nickname].

'/cycle -app' will restart Konversation.

'/cycle -server' closes all tabs of the connection context and
creates a new connection with the same settings as the old one.
Notably this means that it doesn't necessarily rejoin all the
channels you were in previously -- it would be fairly easy to
make it do that, but I'm not sure it should; I'm currently
thinking of /cycle more of starting over fresh. Might change
my mind later though.

'/cycle channel' closes the channel tab and rejoins the channel.

'/cycle nickname' closes a corresponding query and then reopens
it (if there's no such query, it will instead fall through to
showing a help text).

A parameter-less '/cycle' in server, channel and query tabs
does the equivalent of '/cycle -server', '/cycle <that channel>
and '/cycle <that query nick>', respectively.