Bookmark URL generation improvements.
Check whether QUrl accepts the network name as a hostname before
using it, otherwise we get into trouble in the KDE bookmarking
code. Fall back to the server hostname if we have to. Fixes book-
marking when the network name e.g. contains a space.
BUG:255543

Also excise the leading # in standard channel names - we handle
both cases fine, but it's implicit in the IRC url "standard".