Fix a potential crash due to an unchecked cast.
Avoid casting to KAction altogether, which would fail if we're
dealing with a QAction. This is a pretty subtle porting issue,
as KAction didn't use to be a QAction subclass in KDE 3 and
mixing of the two didn't really occur in practice.

Thanks to Aurélien for noticing.
CCMAIL:agateau@kde.org