Expose Konvi's current locale in its own process environment.
The idea is to provide it as KONVERSATION_LANG variable to scripts
we launch, by virtue of the script process inheriting the environ-
ment of the calling process.

Unfortunately the startDetached() calls in KProcess/QProcess don't
support setting up an environment for the process to be started (see
QTBUG-2284), which I would have preferred.