Fix ISON timer initialization.
Adding an initial nick to the WNO list for a network would fail
to kick off the ISON timer if it was done via the context menu
actions or /notify (adding via the config dialog page or the
WNO tab would work). Fixed now.

Also avoiding to pointlessly reschedule the ISON timer every time
the app settings change now, and some more improvements to the
WNO getters in Preferences.