Remove the WNO list controls from the config dialog page.
It's been made redundant by the improved UI of the WNO tab that
supports in-place editing these days. It was originally going to
go away with the merge of the new config dialog, but since that's
not ready and there are bugs in there I don't feel like wasting
time fixing, let's yank it now.