Changed method of remember/marker line tracking.
Qt 4.7.4 introduces a change to the handling of
QTextBlock::userState that makes it unusable for the purposes of
tracking which block represents a marker or remember line. Previous
versions of Qt also changed or otherwise corrupted the userState
value causing the loss of synchronization with the data held in the
IRCView itself. Blocks are now tracked via the userData field, which
was unmolested in the change made to QTextDocument.

Warning: this commit introduces a large volume of debugging output.

For more information about the Qt change, see QTBUG-20916.

BUG: 210106
CCBUG: 260598