Refactor NAMES processing in Channel.
This gets rid of the convoluted list-of-lists approach to
handling NAMES messages and the associated state tracking
members, and instead processes a single list in FIFO mode.

The processing code has also been optimized for the common
case of the nick queue being empty and to avoid requesting
unnecessary nicklist resports, which occurs e.g. on every
nick change or part when processing is kicked off to make
sure the channel is up to date.