Pull the default implementations of appendInputText & Co up into ChatWindow.
The others are isInsertSupported and getTextInLine.

We're finally going to move to a three-tiered class hierarchy of
View:ChatWindow:{Channel,Query,StatusPanel,...} soon to have a common
superclass for the, well, chat windows without impacting the other
views, but until we do, it makes more sense to have this in ChatWindow
than duplicate it all over the place - it had already gotten out of
sync between them, even :/. It'll also mean less work when creating
the new hierarchy.