Port from subprocess.Popen to higher-level functions.
Using check_output() raises the dependency to Python 2.7, which
should be OK for master (media already required 2.7 for the same
reason).

Also adds some error handling to i18n.locale_directories() to make
it fail gracefully, allowing the script to continue to run but
warning visibly about the lack of translations.