Rewrite sysinfo in Python.
Not that much in the way of new features, though it's
a bit more configurable now (emulating the media script
in the approach) and support for reporting the distro
name version either via /etc/os-release or invoking
lsb_release was added. KDE version reporting is also
a bit nicer now, differentiating between running inside
a KDE session and not.