Simplify logic behind the "Show/Hide Konversation" action.
This action is currently implemented such that if the window is not
hidden, not minimised and on the current desktop, the window is hidden,
otherwise the window is shown (and moved to the current desktop if
necessary). This behaviour can be frustrating if the window is not
hidden, not minimised, on the current desktop, and covered by other
windows. Since the window can't be seen, one expects the action to show
the window, but instead it hides it completely.

With this change, the action ignores whether or not the window is on the
current desktop, but instead checks if Konversation is the active
window. This simplifies the behaviour to: If Konversation is the active
window, hide it. Otherwise, show it, move it to the current desktop, and
activate it.

This is also consistent with the behaviour of the equivalent action in
Amarok.

REVIEW:104116