Fix misleading log formatting when /msg'ing from a channel window.
When doing "/msg foo bar" in a channel, we used to log "<foo> bar"
in the channel log. This is very misleading, since it looks as if
'foo' said it rather than sending it *to* foo. This commit changes it
to "<-> foo> bar", which looks a bit odd to me but it's consistent
with what gets shown in the ircview.

This partially fixes the complaints in bug 100369. The other part of
that bug is logging the message in the recipient's log
instead of the channel's, but that's apparently quite hard.

CCBUG: 100369