Make sure the hidden label has been shown before sizeHint() gets called.
Turns out sizeHint() may be called while the label's parent is invisible
(read: the dialog is closed), and if that's the first call to sizeHint()
the label won't have ever been shown and thus not begun adjusting the size
of its document, causing us to return a too-small hint. Because
updateItemWidgets() may run on show before sizeHint() does again this can
lead to ill-layouted item labels, basing their geometry on the broken old
hints.

It also turns out that the label doesn't need to be shown every time we
want it to update its document size; once ever while the parent chain
is shown, too, is enough, so we can get away with doing less work than
before.