Fix showing the away overlay icon in the nicklist What's This.
BUG:301059

Also lays out the text a bit nicer and makes the code backing the
tooltip a little less batshit insane.

It would be nicer to show a blended example, but stuffing the
blended pixmap into a temp file to be able to access it from HTML
isn't worth it. Once more I wish the resource system could register
content dynamically at runtime.