Squelch spurious "Ambiguous Shortcut" warning
SearchBar manages the shortcut for ESC, but does it in a way that
ViewContainer can't account for. If Focus New Tabs is not enabled,
the ViewContainer will reset the action's status without informing
SearchBar.

Now ViewContainer will send an event through the event-filter managing
the shortcut so that it can deactivate the aciton, avoiding the error
popup.

BUG:314433