Flush inbound queues when flushing outbound queues
If a quit message comes through at the same time as the socket is
closed by the server, a mix of quit and reconnect actions occur.
This is caused by the return to the event loop between pulling the
data out of the socket and actually processing it. This would close
all channels, for instance, because our nick is still in the nicklist
when the quit message goes through the inputfilter.