Ignore context ChatWindow's notification setting in NotificationHandler::nick(On|Off)line.
Watched Nicks coming online or offline is connection-global, not
specific to any particular ChatWindow, so checking whether the one
passed as context (in practice the call sites use the StatusPanel,
and the notification handler only cares to get the corresponding
Server) has notifications enabled or not makes no sense here.

As "parking" on a status tab before switching away from the app is
a common behavior, as is switching off its notifications to avoid
certain highlights during connect, this is the likely cause of in-
termittent reports of this KNotify event not working reliably.