Get back the tree view in the config dialog.
Unfortunately this (re-, actually) introduces a fork of
KConfigDialog modified not to be hell-bent on hiding the
ability of KPageDialog to do trees.

Long-term we actually want to get rid of that three, but
for now this is the least offensive way to get it back.
Which is saying a lot.