Japanese charsets fixes.
We used to override QTextCodec::codecForLocale() by explicitly
defaulting to jis7 for jp_JP, but KCharsets/Qt5 seem to no longer
support jis7. We also used to mangle ISO-2022-JP to jis7 - which
actually is supported. So lets default to ISO-2022-JP and mangle
jis7 to ISO 2022-JP instead to keep existing configs working ...

However this showed up as a crash due to not checking for null
codecs, so let's add a check too.

BUG:362401