More performance work
* Reverse text view vertical layout again, was a crude hack for making
  initial messages come up from the bottom of the view.
* Therefore switch from container::prepend to container::append again
* Switch from QList to QVector, it's much faster at QString appends.

Batch allocation using reserve() is kept.

This kind of exhausts all the easy things to do. To speed up further,
more stuff in the message delegate could be moved into C++ and cached,
e.g. the nickname abbreviations shown in the fake avatars, at the cost
of theming flexibility. It may not actually be necessary though,
performance is quite good now.