Force QT_QUICK_CONTROLS_STYLE=org.kde.desktop outside Plasma
Adds a runtime dep on qqc2-style-desktop. The Flatpak recipe
has been updated to pull this in.

qqc2-style-desktop is destined to be part of KDE Frameworks 5
soon, then we can depend on it via CMake. In the meantime the
build instructions for the branch on the wiki have been updated
to reflect the dep.

Instead of qputenv we should be using the QQuickStyle C++ API,
but Kirigami doesn't read this yet.
CCMAIL:notmart@gmail.com