Get rid of window nesting
We now hide the QWidget-based main window at startup and show the
Qt Quick window non-embedded. F10 or a button in Settings mode
reveals the legacy window; it can be closed and reopened at will.

There's a bunch of hard-coded shortcuts now to open dialogs (e.g.
F2 for the Server List) but most shortcuts are broken for now as
they relied on KXmlGuiMainWindow before. I might write a stop-gap
bridge for this next.

The reason for getting rid of embedding (and hence toggling) for
now is that it breaks touch handling on Wayland and Android
completely, along with a bunch of focus issues we've been plagued
with. And long-term we want to get rid of QWidget anyway.