Allow channel history to be deleted
Summary:
Any channel which has ever been attempted to be joined is currently
saved to history, including unsuccessful attempts and typos. These
can be seen in the join channel dialog (Ctrl+J).

The only way previously for a channel to be deleted from history
was to manually edit ~/.config/konversationrc

This is now rectified, with a delete button in the join channel
dialog, enabling users to delete previously selected channels from
history.

Reviewers: #konversation, psn

Reviewed By: #konversation, psn

Subscribers: #konversation

Tags: #konversation

Differential Revision: https://phabricator.kde.org/D28644