Channel Settings UI: bump max value of "User limit" to 99999
Max value was the default 99, which would result in...
a) any higher value read from the channel mode being reset to 99
b) not allow users to set higher values if wanted.

As applying the channel settings from the dialog always picks all the
values of the dialog, also those not edited, in case of a) this also
would mean changing the user limit in the channel to 99 then.

RFCs are not known to define a max user limit, so this is server defined.
While there is no path explored to query any max user limit capabilities,
and servers implementations seems to just map too high numbers to random
possible ones, there is no well defined max value to be used for the
dialog.

Using 99999 as arbitrary value should serve for now all expected use-cases.
To be raised later if there is respective feedback.

FIXED-IN: 1.8.21121
BUG: 446673