Port away from KIdleTime::simulateUserActivity()
The goal of the existing method simulateUserActivity was to make
konversation react as though the user has interacted with the system
resetting away states and restarting timers.

`KIdleTime::simulateUserActivity()` tells *every* application that the
user has interacted with the system so they can reset the timers.

It's unnecesary for konversation to do this. This method will be
deprecated in newer KIdleTime.