Add error handling and modernize kpart loading
The original code included quite a bit of error handling, this has been
preserved and consolidated so that a meaningful error message could be
displayed.

Given the string freeze and the unlikelihood of these errors, the
implementation of proper messages is future work. For now an error is
left as a number.