Provide scripts with an env var to find qdbus
While the qdbus tool has been a RUNTIME dependency of Konversation since 2008,
most distros didn't add a requirement. Given there's no standard way to
advertise runtime deps, qdbus is now a build dep in addition to being a
runtime dep.

Unfortunately qdbus doesn't have a consistent name nor location. This attempts
to find it by asking Qt where the binaries probably got installed and then
tries a few possible names.

A script run from inside Konversation now expects to receive an environment
variable "KONVERSATION_DBUS_BIN" that is the full pathname of the actual
program. This is set by the script launcher if it isn't present in the
environment on launch.

BUG: 349675
FIXED-IN: 1.10.24020