Set tray icon to active when the window is closed
Some tray implementations will only show "Active" or "Needs Attention"
tray icons, and offer no means of configuration. The tray icon will now
be set to "Active" when the window is closed and the tray icon is
enabled, which will also cause Plasma's tray to show the icon when it
is set to "Shown when relevant".

BUG:494194